requirements.txt
streamlit==1.44.1
python-dotenv==1.0.1
mysql-connector-python==9.2.0
passlib[bcrypt]==1.7.4
openai>=1.0.0
huggingface-hub==0.30.2
youtube-transcript-api==0.6.3
requests==2.32.3
beautifulsoup4==4.13.3
pypdf==5.4.0
python-docx==1.1.2
python-pptx==1.0.2
email-validator==2.2.0
pandas==2.2.3
reportlab==4.3.1
gTTS==2.5.4
youtube-search-python==1.6.6
Pillow>=10.0.0
deep-translator>=1.11.0
SpeechRecognition>=3.10.0
audio-recorder-streamlit>=0.0.8
fpdf2>=2.7.0
yagmail>=0.15.0
pydub>=0.25.1
